create a separate text input tag picker
